[DEV-5856] renaming files for consistency
